Update zmnotify.py
This fixes a two layer bug. If the app initializes wit the occupied bool set to true (in other-words I am home) the state of the bool is grabbed as a string 'on' rather than True or False. Further, no check is performed to see if I am home so it defaults to unoccupied until I toggle the bool in HA. So I added a simple if statement to fix that.